Rewrite mouse cloning
Adding support for gnome-shell 46 has introduced a new bug:

When cursor cloning is enabled, resizing windows causes the cursor to
freeze up until ESC pressed, after which clicking on windows to change
focus no longer works correctly until the extention is disabled.

There did not appear to be an easy "fix" for this, as there aren't any
obvious differences in how cursors are dealt with in between gnome-shell
45 and 46.

This commit simplifies and refactors mouse cloning, fixing the bug in
the process.

Note that it does not jive with the magnification tool -- it ends up
showing two cursors in slightly different positions.  Adjust the warning
in the preferences window to read as such, unfortunately breaking
translations in po/.

See GitHub issue:
https://github.com/jkitching/soft-brightness-plus/issues/20